added dev tasks to install and run redis server

+++ lib/tasks/develop.rake
@@ -1,5 +1,6 @@
 require 'open-uri'
 download_dirpath = File.join( File.dirname( __FILE__ ), 'download' )
+opt_dirpath = File.join( File.dirname( __FILE__ ), 'opt' )
 redis_version = '2.8.3'
 
 namespace :development do
@@ -9,14 +10,37 @@ namespace :development do
     Dir.mkdir download_dirpath unless Dir.exists?( download_dirpath )
   end
   
+  desc "ensure opt directory exists" 
+  task :ensure_opt_directory do
+    Dir.mkdir opt_dirpath unless Dir.exists?( opt_dirpath )
+  end
+  
   desc "download redis source"
   task :download_redis_source => :ensure_download_directory do
-    print "Downloading redis source..."
-    Dir.chdir download_dirpath
-    open( "redis-#{redis_version}.tar.gz", 'wb' ) do |file|
-      file << open("http://download.redis.io/releases/redis-#{redis_version}.tar.gz").read
+    if File.exists?( File.join( download_dirpath, "redis-#{redis_version}.tar.gz" ) )
+      puts "redis source tarball already downloaded...skipping (remove download/redis-#{redis_version}.tar.gz and re-run to force)"
+    else
+      print "Downloading redis source..."
+      Dir.chdir download_dirpath
+      open( "redis-#{redis_version}.tar.gz", 'wb' ) do |file|
+        file << open("http://download.redis.io/releases/redis-#{redis_version}.tar.gz").read
+      end
+      puts "done."
     end
-    puts "done."
+  end
+  
+  desc "install temporary redis installation"
+  task :install_temp_redis => [ :ensure_opt_directory, :download_redis_source ] do
+    Dir.chdir( opt_dirpath )
+    system "tar xzf ../download/redis-#{redis_version}.tar.gz"
+    Dir.chdir( Dir.join( opt_dirpath, "redis-#{redis_version}" ) )
+    system "make"
+  end
+  
+  desc "start temporary redis service"
+  task :start_temp_redis do
+    Dir.chdir( Dir.join( opt_dirpath, "redis-#{redis_version}" ) )
+    system "src/redis-server"
   end
   
 end
